자바 실습 제32장 ~ 제33장

+++ Java/chapter02/src/OperatePromotion.java
@@ -4,7 +4,7 @@ public class OperatePromotion {
 	public static void main(String[] args) {
 		short num1 = 11;
 		short num2 = 22;
-		short num3 = num1 + num2;
+		int num3 = num1 + num2;
 		System.out.println(num3);
 	}
